requirements.txt
altair==5.5.0
annotated-types==0.7.0
attrs==25.3.0
bandit==1.8.6
blinker==1.9.0
blis==1.3.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
confection==0.1.5
cryptography==45.0.5
cymem==2.0.11
filelock==3.18.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.71.2
httplib2==0.22.0
idna==3.10
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
langcodes==3.5.0
language_data==1.3.0
lxml==6.0.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.13
mypy==1.17.0
mypy_extensions==1.1.0
narwhals==1.47.1
numpy==2.3.1
packaging==25.0
pandas==2.3.1
pathspec==0.12.1
pbr==6.1.1
pdfminer.six==20250506
pdfplumber==0.11.7
phonenumbers==8.13.55
pillow==11.3.0
preshed==3.0.10
presidio-analyzer==2.2.35
presidio-anonymizer==2.2.35
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.1.1
python-magic==0.4.27
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-file==2.1.0
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
ruff==0.12.4
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
smmap==5.0.2
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stevedore==5.4.1
streamlit==1.47.0
tenacity==9.1.2
thinc==8.3.6
tldextract==5.3.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typer==0.16.0
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.1
wrapt==1.17.2
